feat(contributes): add start()
