--- frameworks/contributes/ts/index.spec.ts
@@ -1,6 +1,0 @@
-import { commands, keyBindings } from '.'
-
-test('exports', () => {
-  expect(commands).toBeDefined()
-  expect(keyBindings).toBeDefined()
-})
