+++ frameworks/contributes/ts/module.spec.ts
@@ -1,0 +1,9 @@
+import { start } from './module'
+
+describe('start()', () => {
+  test('returns commands', async () => {
+    const { commands, keyBindings } = await start({ commands: {}, keyBindings: {} })
+    expect(commands).toBeDefined()
+    expect(keyBindings).toBeDefined()
+  })
+})
