+++ frameworks/contributes/ts/types.ts
@@ -1,6 +1,3 @@
-import { createStore } from '@just-web/states'
-import { record } from 'type-plus'
-
 export interface Command {
   /**
    * The command id. e.g. `just-web.showCommandPalette`
@@ -20,4 +17,17 @@ export interface Command {
   description?: string
 }
 
-export const commands = createStore(record<string, Command>())
+export interface KeyBinding {
+  /**
+   * Command to bind to.
+   */
+  command: string,
+  /**
+   * Default key
+   */
+  key?: string,
+  /**
+   * MacOS specific key
+   */
+  mac?: string
+}
